requirements.txt
aiodns==3.0.0
aiogram==2.14.3
aiohttp==3.7.4.post0
aioredis==1.3.1
async-timeout==3.0.1
asyncpg==0.24.0
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.9.3
brotlipy==0.7.0
bs4==0.0.1
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
Faker==8.12.0
hiredis==2.0.0
idna==3.2
lxml==4.6.3
maxminddb==2.0.3
maxminddb-geolite2==2018.703
mimesis==4.1.3
multidict==5.1.0
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.1
pycares==4.0.0
pycparser==2.20
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
text-unidecode==1.3
typing-extensions==3.10.0.0
ujson==4.1.0
uvloop==0.16.0
vt-py==0.7.2
yarl==1.6.3
exifread==2.3.2
